chore: move watch communication to dispatch vc
This will help in communicating logged in / out since
the FeedingVM has no idea about logged in state
